backend/pyproject.toml
[project]
name = "firesight-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosmtplib>=5.1.0",
    "fastapi>=0.135.2",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.21.0",
    "uvicorn[standard]>=0.42.0",
]
